Organize App files
