+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/index.js":"1","/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/App.js":"2","/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/reportWebVitals.js":"3","/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/movieData.js":"4"},{"size":500,"mtime":1612305731526,"results":"5","hashOfConfig":"6"},{"size":287,"mtime":1612396592584,"results":"7","hashOfConfig":"6"},{"size":362,"mtime":1612305731527,"results":"8","hashOfConfig":"6"},{"size":12968,"mtime":1612396508543,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"1y57pmc",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/index.js",[],["19","20"],"/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/App.js",[],"/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/reportWebVitals.js",[],"/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/movieData.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/index.js":"1","/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/reportWebVitals.js":"2","/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/App/App.js":"3","/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/Data/movieData.js":"4"},{"size":389,"mtime":1612396896228,"results":"5","hashOfConfig":"6"},{"size":362,"mtime":1612305731527,"results":"7","hashOfConfig":"6"},{"size":293,"mtime":1612396758138,"results":"8","hashOfConfig":"6"},{"size":12968,"mtime":1612396508543,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1y57pmc",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/index.js",[],"/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/reportWebVitals.js",[],["19","20"],"/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/App/App.js",[],"/Users/christinacook/turing/mod3/projects/proj-1/rancid-tomatillos/src/Data/movieData.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
